+++ runway/__init__.py
@@ -1,2 +1,2 @@
 """Set package version."""
-__version__ = '0.47.2'
+__version__ = '0.47.3'